`get_free_tcp_port` and `is_port_available` move to utils.py
